# docs/openapi.yaml
openapi: 3.0.4
info:
  title: The Crazy Unicorn Shoppe
  version: 1.0.0
  description: >-
    This is an e-commerce full-stack app made by Steve Majewski for practice,
    the content is intended to be silly and far from serious, but it will be
    fully functional as far as creating accounts and logging in with hashed
    passwords. The data will be stored in a Postgres Database.
servers:
  - url: www.localhost:4700
    variables: {}
...
paths:
...
  /order:
    summary: handles order related operations
    post:
      summary: creates an order
      description: >-
        This operation takes in JSON data in the request body and returns a new
        order id.
      operationId: post_order
      requestBody:
        description: Info to create a new order.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/New_order'
      responses:
        '200':
          description: OK
          content:
            application/json:
              examples:
                create order success:
                  value: |-
                    {
                      "order_id": 47,
                      "msg": "order created"
                    }
        '401':
          description: not authorized
          content:
            text/plain; charset=utf-8:
              examples:
                not authorized:
                  value: You must be logged in to perform this action.                    
        '500':
          description: internal server error
          content:
            text/plain; charset=utf-8:
              examples:
                server error:
                  value: Something went wrong on our end, please try again.
      tags:
        - orders
  /order/{id}:
    summary: handles operations for a specific order
    get:
      summary: get order info
      description: >-
        This will return info in JSON form about the order id that was included
        as a url parameter.
      operationId: get_order
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the order whose info we seek.
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              examples:
                get order info success:
                  value: |-
                    {
                      "order_id": 47,
                      "user_id": 9,
                      "cart_id": 3,
                      "order_date": "2025-06-13 11:10:17.046068-04",
                      "order_total": 75,
                      "order_status": "pending",
                      "payment_method": "credit card",
                      "free_shipping_elligible": false,
                      "order_items": []
                    }
        '401':
          description: not authorized
          content:
            text/plain; charset=utf-8:
              examples:
                not authorized:
                  value: You must be logged in to perform this action.
        '404':
          description: not found
          content:
            text/plain; charset=utf-8:
              examples:
                not found:
                  value: We could not find a order with that id.
        '500':
          description: internal server error
          content:
            text/plain; charset=utf-8:
              examples:
                server error:
                  value: Something went wrong on our end, please try again.
      tags:
        - orders
    put:
      summary: updates order status
      description: >-
        This operation will take in a new status in the request body and update
        the database. Choices are pending, complete, canceled
      operationId: put_order
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the order whose status we seek to change.
          schema:
            type: string
      requestBody:
        description: Updates an order status.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Update_order_status'
      responses:
        '200':
          description: OK
          content:
            application/json:
              examples:
                order status update success:
                  value: |-
                    {
                      "msg": "Order status has been updated",
                      "order_status": "complete" 
                    }
        '401':
          description: not authorized
          content:
            text/plain; charset=utf-8:
              examples:
                not authorized:
                  value: You must be logged in to perform this action.
        '404':
          description: not found
          content:
            text/plain; charset=utf-8:
              examples:
                not found:
                  value: We could not find a order with that id.
        '500':
          description: internal server error
          content:
            text/plain; charset=utf-8:
              examples:
                server error:
                  value: Something went wrong on our end, please try again.
      tags:
        - orders
    delete:
      summary: cancel an order
      description: This operation will cancel an order.
      operationId: delete_order
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the order we seek to cancel.
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              examples:
                order canceled success:
                  value: |-
                    {
                      "msg": "The order has been successfully canceled.",
                      "order_status": "canceled" 
                    }
        '401':
          description: not authorized
          content:
            text/plain; charset=utf-8:
              examples:
                not authorized:
                  value: You must be logged in to perform this action.
        '404':
          description: not found
          content:
            text/plain; charset=utf-8:
              examples:
                not found:
                  value: We could not find a order with that id.
        '500':
          description: internal server error
          content:
            text/plain; charset=utf-8:
              examples:
                server error:
                  value: Something went wrong on our end, please try again.
      tags:
        - orders
components:
  schemas:
...
    New_order:
      type: object
      properties:
        user_id:
          type: integer
          format: int32
          example: 47
        cart_id:
          type: integer
          format: int32
          example: 225
        order_total:
          type: integer
          format: int32
          example: 500
        payment_method:
          type: string
          example: paypal
    Update_order_status:
      type: object
      properties:
        updatedOrderStatus:
          type: string
          example: complete